prevent beeps when navigating multi-line commands
`man zshzle` at the "zle widget" section:

A non-zero return status causes the shell to beep when the widget exits,
unless the BEEP options was unset or the widget was called via the zle
command. Thus if a user defined widget requires an immediate beep, it
should call the beep widget directly.

https://github.com/robbyrussell/oh-my-zsh/pull/215#issuecomment-1328406