Fix logic used to build fuzzy search regex
The intent was to insert '*' between all characters in the query, except
for those which were escaped with '\'. What the old implementation did,
however, was strip out all backslashes and place '*' between all
remaining characters, as when a backslash was detected the loop would
skip to the next character without appending the backslash to the regex.

The new logic is something like "add the character, then add '*' if the
character was not a backslash", which should be correct